Use auto props
