+++ src/Workspaces/Core/Portable/Rename/RenameLocations.SearchResult.cs
@@ -1,0 +1,29 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Collections.Immutable;
+using Microsoft.CodeAnalysis.FindSymbols;
+
+namespace Microsoft.CodeAnalysis.Rename
+{
+    internal sealed partial class RenameLocations
+    {
+        private class SearchResult
+        {
+            public readonly ImmutableHashSet<RenameLocation> Locations;
+            public readonly ImmutableArray<ReferenceLocation> ImplicitLocations;
+            public readonly ImmutableArray<ISymbol> ReferencedSymbols;
+
+            public SearchResult(
+                ImmutableHashSet<RenameLocation> locations,
+                ImmutableArray<ReferenceLocation> implicitLocations,
+                ImmutableArray<ISymbol> referencedSymbols)
+            {
+                this.Locations = locations;
+                this.ImplicitLocations = implicitLocations;
+                this.ReferencedSymbols = referencedSymbols;
+            }
+        }
+    }
+}
